backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.2
microsoft-kiota-abstractions==1.4.6
microsoft-kiota-authentication-azure==1.4.6
microsoft-kiota-http==1.4.6
msal==1.31.0
msal-extensions==1.2.0
msgraph-core==1.1.6
multidict==6.1.0
openai==1.52.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
portalocker==2.10.1
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pywin32==308
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlalchemy-utils==0.41.2
starlette==0.41.2
std-uritemplate==1.0.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2
yarl==1.15.5
zipp==3.20.2
alembic  # for database migrations